test: 구급대원 테스트 완료

+++ backend/member-service/src/main/java/com/codesmith/goojangmember/member/persistence/domain/SafetyCenter.java
@@ -13,14 +13,14 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 @EntityListeners(AuditingEntityListener.class)
 public class SafetyCenter {
     @Id
-    private Integer id;
+    private Long id;
     private String region;
     private String name;
     private String address;
     private String telephone;
     private String fax;
 
-    public SafetyCenter(Integer id, String region, String name, String address, String telephone, String fax) {
+    public SafetyCenter(Long id, String region, String name, String address, String telephone, String fax) {
         this.id = id;
         this.region = region;
         this.name = name;
